% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_cbi_pay_stats.R
\name{cbi_read_data}
\alias{cbi_read_data}
\title{Read and Clean Card Payment Statistics}
\usage{
cbi_read_data(file_path)
}
\arguments{
\item{file_path}{The path to the Excel file containing the data.}
}
\value{
A clean tibble with the payment statistics.
}
\description{
This function reads the "Master Data" sheet from the monthly card payments
Excel file, cleans the column names, and formats the reporting period.
}
