% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonthlyPayStatsForecast-package.R
\docType{package}
\name{MonthlyPayStatsForecast-package}
\alias{MonthlyPayStatsForecast}
\alias{MonthlyPayStatsForecast-package}
\title{MonthlyPayStatsForecast: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichaelSandilands/MonthlyPayStatsForecast}
  \item \url{https://michaelsandilands.github.io/MonthlyPayStatsForecast/}
  \item Report bugs at \url{https://github.com/MichaelSandilands/MonthlyPayStatsForecast/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
